requirements.txt
Django==2.2
djangorestframework==3.9.2
pytz==2019.1
sqlparse==0.3.0
django_select2
django_admin
Pillow==6.0.0
ply==3.11
pyparsing==2.4.0
python-thumbnails==0.5.1
django-computed-property==0.2.4
